Seperate stack web handler into own file
