+++ backend/internal/web/handler.go
@@ -26,50 +26,6 @@ type pageData struct {
 	Session     commonauth.Session
 }
 
-type papersListData struct {
-	Items         []paperDomain.Paper
-	Total         int
-	Page          int
-	PageSize      int
-	HasNext       bool
-	PrevPage      int
-	NextPage      int
-	SearchOptions paperDomain.SearchOptions
-	Pagination    []PaginationItem
-}
-
-type stacksListData struct {
-	Items         []stackDomain.Stack
-	Total         int
-	Page          int
-	PageSize      int
-	HasNext       bool
-	PrevPage      int
-	NextPage      int
-	SearchOptions stackDomain.SearchOptions
-	Pagination    []PaginationItem
-}
-
-func NewStacksListData(result stackDomain.SearchResult, opts stackDomain.SearchOptions) stacksListData {
-	return stacksListData{
-		Items:         result.Items,
-		Total:         result.Total,
-		Page:          result.Page,
-		PageSize:      result.PageSize,
-		HasNext:       result.HasNext,
-		PrevPage:      result.Page - 1,
-		NextPage:      result.Page + 1,
-		SearchOptions: opts,
-		Pagination:    BuildPagination(result.Total, result.PageSize, result.Page),
-	}
-}
-
-type stacksPageData struct {
-	pageData
-	StacksCountTotal  int
-	StacksCountPublic int
-}
-
 type alertData struct {
 	Message string
 }
@@ -129,80 +85,6 @@ func handleSidebarStacks(
 	})
 }
 
-func handleStacksPage(
-	logger *slog.Logger,
-	tmpl *template.Template,
-	hankoAPIURL string,
-	stackService *stackApp.StackService,
-) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		w.Header().Set("Content-Type", "text/html; charset=utf-8")
-
-		session, ok := commonauth.SessionFromContext(r.Context())
-		if !ok {
-			session = &commonauth.Session{}
-		}
-
-		counter, err := stackService.CountPublic(r.Context())
-		if err != nil {
-			logger.Error("count public stacks", "error", err.Error())
-			http.Error(w, http.StatusText(http.StatusInternalServerError), http.StatusInternalServerError)
-			return
-		}
-
-		data := stacksPageData{
-			pageData: pageData{
-				AppVersion:  build.Version,
-				AppGitHash:  build.GitHash,
-				PageName:    pageNameFromPath(r.URL.Path),
-				HankoAPIURL: hankoAPIURL,
-				Session:     *session,
-			},
-			StacksCountTotal:  0,
-			StacksCountPublic: counter,
-		}
-
-		renderTemplate(w, r, tmpl, data)
-	})
-}
-
-func handleStacksMyPage(
-	logger *slog.Logger,
-	tmpl *template.Template,
-	hankoAPIURL string,
-	stackService *stackApp.StackService,
-) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		w.Header().Set("Content-Type", "text/html; charset=utf-8")
-
-		session, ok := commonauth.SessionFromContext(r.Context())
-		if !ok {
-			session = &commonauth.Session{}
-		}
-
-		counter, err := stackService.CountPublic(r.Context())
-		if err != nil {
-			logger.Error("count public stacks", "error", err.Error())
-			http.Error(w, http.StatusText(http.StatusInternalServerError), http.StatusInternalServerError)
-			return
-		}
-
-		data := stacksPageData{
-			pageData: pageData{
-				AppVersion:  build.Version,
-				AppGitHash:  build.GitHash,
-				PageName:    pageNameFromPath(r.URL.Path),
-				HankoAPIURL: hankoAPIURL,
-				Session:     *session,
-			},
-			StacksCountTotal:  0,
-			StacksCountPublic: counter,
-		}
-
-		renderTemplate(w, r, tmpl, data)
-	})
-}
-
 func renderTemplate(w http.ResponseWriter, r *http.Request, tmpl *template.Template, data any) {
 	templateName := "base"
 	if r.Header.Get("HX-Request") == "true" {
@@ -258,90 +140,6 @@ func handlePapersSearch(
 		templateName := "papers/partials/papers-list"
 		if err := tmpl.ExecuteTemplate(w, templateName, data); err != nil {
 			http.Error(w, http.StatusText(http.StatusInternalServerError), http.StatusInternalServerError)
-		}
-	})
-}
-
-func handleStacksSearchPublic(
-	logger *slog.Logger,
-	tmpl *template.Template,
-	stackService *stackApp.StackService,
-) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		w.Header().Set("Content-Type", "text/html; charset=utf-8")
-
-		opts := searchOptionsFromRequest(r)
-		result, err := stackService.Search(r.Context(), opts)
-		if err != nil {
-			logger.Error("read stacks", "error", err.Error())
-			http.Error(w, err.Error(), http.StatusInternalServerError)
-			return
-		}
-
-		data := NewStacksListData(result, opts)
-
-		templateName := "stacks/partials/stacks-list"
-		if err := tmpl.ExecuteTemplate(w, templateName, data); err != nil {
-			http.Error(w, http.StatusText(http.StatusInternalServerError), http.StatusInternalServerError)
-			return
-		}
-	})
-}
-
-func handleStacksSearchByOwner(
-	logger *slog.Logger,
-	tmpl *template.Template,
-	stackService *stackApp.StackService,
-) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		w.Header().Set("Content-Type", "text/html; charset=utf-8")
-
-		session, ok := commonauth.SessionFromContext(r.Context())
-		if !ok {
-			session = &commonauth.Session{}
-		}
-
-		opts := searchOptionsFromRequest(r)
-		result, err := stackService.SearchByOwner(r.Context(), session.UserID, opts)
-		if err != nil {
-			logger.Error("read stacks", "error", err.Error())
-			http.Error(w, err.Error(), http.StatusInternalServerError)
-			return
-		}
-
-		data := NewStacksListData(result, opts)
-		templateName := "stacks/partials/stacks-list"
-		if err := tmpl.ExecuteTemplate(w, templateName, data); err != nil {
-			http.Error(w, http.StatusText(http.StatusInternalServerError), http.StatusInternalServerError)
-			return
-		}
-	})
-}
-
-func handleStacksStatsByOwner(
-	logger *slog.Logger,
-	tmpl *template.Template,
-	stackService *stackApp.StackService,
-) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		w.Header().Set("Content-Type", "text/html; charset=utf-8")
-		ctx := r.Context()
-
-		session, ok := commonauth.SessionFromContext(ctx)
-		if !ok {
-			session = &commonauth.Session{}
-		}
-		stats, err := stackService.GetStatsByOwner(ctx, session.UserID)
-		if err != nil {
-			logger.Error("read stacks", "error", err.Error())
-			http.Error(w, err.Error(), http.StatusInternalServerError)
-			return
-		}
-
-		templateName := "stacks/partials/stats-my"
-		if err := tmpl.ExecuteTemplate(w, templateName, stats); err != nil {
-			http.Error(w, http.StatusText(http.StatusInternalServerError), http.StatusInternalServerError)
-			return
 		}
 	})
 }
